+++ settings.gradle.kts
@@ -7,4 +7,4 @@
  * in the user manual at https://docs.gradle.org/6.1.1/userguide/multi_project_builds.html
  */
 
-rootProject.name = "veryFit2googleFit"
+rootProject.name = "vFit2gFit"
