+++ src/main/kotlin/io/github/durun/vFit2gFit/App.kt
@@ -1,0 +1,5 @@
+package io.github.durun.vFit2gFit
+
+import io.github.durun.vFit2gFit.app.AppCommand
+
+fun main(args: Array<String>) = AppCommand.main(args)
