+++ src/main/kotlin/io/github/durun/vFit2gFit/reader/HeartRateReader.kt
@@ -1,0 +1,2 @@
+package io.github.durun.vFit2gFit.reader
+
