+++ src/main/kotlin/io/github/durun/vFit2gFit/writer/db/table/HeartRateTable.kt
@@ -1,4 +1,4 @@
-package veryFit2googleFit.writer.db.table
+package io.github.durun.vFit2gFit.writer.db.table
 
 import org.jetbrains.exposed.sql.Column
 import org.jetbrains.exposed.sql.Table
