--- src/main/kotlin/veryFit2googleFit/App.kt
@@ -1,5 +1,0 @@
-package veryFit2googleFit
-
-import veryFit2googleFit.app.AppCommand
-
-fun main(args: Array<String>) = AppCommand.main(args)
