--- src/main/kotlin/veryFit2googleFit/reader/HeartRateReader.kt
@@ -1,2 +1,0 @@
-package veryFit2googleFit.reader
-
