+++ src/test/kotlin/io/github/durun/vFit2gFit/AppTest.kt
@@ -1,7 +1,7 @@
 /*
  * This Kotlin source file was generated by the Gradle 'init' task.
  */
-package veryFit2googleFit
+package io.github.durun.vFit2gFit
 
 import kotlin.test.Test
 import kotlin.test.assertNotNull
